Refactored shader functions to separate file
Changed vectors from doubles to floats

+++ src/headers/vec.h
@@ -24,12 +24,12 @@ typedef struct {
 } Ray;
 typedef Ray* RayRef;
 
-static inline double deg2rad(double deg) {
-    return deg * (PI / 180.0);
+static inline float deg2rad(float deg) {
+    return (float)(deg * (PI / 180.0));
 }
 
-static inline double vec_mag(Vector a) {
-    return sqrt(a.x*a.x + a.y*a.y + a.z*a.z);
+static inline float vec_mag(Vector a) {
+    return (float)sqrt(a.x*a.x + a.y*a.y + a.z*a.z);
 }
 
 static inline Vector vec_add(Vector a, Vector b) {
@@ -48,7 +48,7 @@ static inline Vector vec_sub(Vector a, Vector b) {
     };
 }
 
-static inline Vector vec_scale(Vector a, double scale) {
+static inline Vector vec_scale(Vector a, float scale) {
     return (Vector) {
             .x = a.x * scale,
             .y = a.y * scale,
@@ -57,7 +57,7 @@ static inline Vector vec_scale(Vector a, double scale) {
 }
 
 static inline Vector vec_unit(Vector a) {
-    double mag = vec_mag(a);
+    float mag = vec_mag(a);
 
     // prevent division by zero
     if (mag == 0)
@@ -66,11 +66,11 @@ static inline Vector vec_unit(Vector a) {
     return vec_scale(a, 1.0 / mag);
 }
 
-static inline double vec_dot(Vector a, Vector b) {
+static inline float vec_dot(Vector a, Vector b) {
     return a.x*b.x + a.y*b.y + a.z*b.z;
 }
 
-static inline double vec_angle(Vector a, Vector b) {
+static inline float vec_angle(Vector a, Vector b) {
     return acos(vec_dot(a, b));
 }
 
@@ -98,21 +98,21 @@ static inline Vector vec_rot(Vector point, Vector angle) {
     };
 
     v1 = (Vector) {
-            .x = point.x * cos(degAngle.z) - point.y * sin(degAngle.z),
-            .y = point.x * sin(degAngle.z) + point.y * cos(degAngle.z),
-            .z = point.z
+            .x = (float)(point.x * cos(degAngle.z) - point.y * sin(degAngle.z)),
+            .y = (float)(point.x * sin(degAngle.z) + point.y * cos(degAngle.z)),
+            .z = (float)(point.z)
     };
 
     v2 = (Vector) {
-            .z = v1.z * cos(degAngle.y) - v1.x * sin(degAngle.y),
-            .x = v1.z * sin(degAngle.y) + v1.x * cos(degAngle.y),
-            .y = v1.y
+            .z = (float)(v1.z * cos(degAngle.y) - v1.x * sin(degAngle.y)),
+            .x = (float)(v1.z * sin(degAngle.y) + v1.x * cos(degAngle.y)),
+            .y = (float)(v1.y)
     };
 
     v3 = (Vector) {
-            .y = v2.y * cos(degAngle.x) - v2.z * sin(degAngle.x),
-            .z = v2.y * sin(degAngle.x) + v2.z * cos(degAngle.x),
-            .x = v2.x
+            .y = (float)(v2.y * cos(degAngle.x) - v2.z * sin(degAngle.x)),
+            .z = (float)(v2.y * sin(degAngle.x) + v2.z * cos(degAngle.x)),
+            .x = (float)(v2.x)
     };
 
     return v3;
@@ -120,7 +120,7 @@ static inline Vector vec_rot(Vector point, Vector angle) {
 }
 
 static inline Vector vec_reflect(Vector vec, Vector axis) {
-    double scale = 2 * vec_dot(axis, vec);
+    float scale = 2 * vec_dot(axis, vec);
     return vec_unit(vec_sub(vec, vec_scale(axis, scale)));
 }
